'default'

+++ app/Jobs/UpdateRecentsJob.php
@@ -2,7 +2,13 @@
 
 namespace App\Jobs;
 
+use App\User;
+use App\Model\Liste;
+use App\Model\Recent_list;
+use App\Model\Recent_user;
 use App\Model\Recent_movie;
+use App\Model\Recent_person;
+use App\Model\Recent_series;
 use Illuminate\Bus\Queueable;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Queue\InteractsWithQueue;
@@ -51,11 +57,15 @@ class UpdateRecentsJob implements ShouldQueue
             $keep = Recent_person::where('user_id', $this->userId)->latest('updated_at')->take(5/* config('constants.recently_viewed.last_n') */)->pluck('id');
             Recent_person::where('user_id', $this->userId)->whereNotIn('id', $keep)->delete();
         } else if($this->type === 'user') {
+            $user = User::where(['id' => $request->id]);
+            if(!$user->count() > 0) return;
             $recent = Recent_user::updateOrCreate(array('user_id' => $this->userId, 'subject_id' => $this->objId));
             $recent->touch();
             $keep = Recent_user::where('user_id', $this->userId)->latest('updated_at')->take(5/* config('constants.recently_viewed.last_n') */)->pluck('id');
             Recent_user::where('user_id', $this->userId)->whereNotIn('id', $keep)->delete();
         } else if($this->type === 'list') {
+            $list = Liste::where(['id' => $request->id]);
+            if(!$list->count() > 0) return;
             $recent = Recent_list::updateOrCreate(array('user_id' => $this->userId, 'list_id' => $this->objId));
             $recent->touch();
             $keep = Recent_list::where('user_id', $this->userId)->latest('updated_at')->take(5/* config('constants.recently_viewed.last_n') */)->pluck('id');
